syscall package functions that have different signatures on different platforms. Split the code into platform-specific files
